Update README.md with DeepL extension details
The code changes in this commit update the README.md file to provide a
concise description of the DeepcastX extension. Additionally, it
clarifies that users who have a DeepL API should use Deepcast for
translations in Raycast. The commit also removes outdated installation
instructions and adds information on how to obtain the DeepLX API URL.